# dist/swagger/swagger.json
{
    "components": {
...
        "schemas": {
            "FlattenMaps_T_": {
                "properties": {},
                "type": "object"
            },
            "Document_any.any.any_": {
                "$ref": "#/components/schemas/FlattenMaps_T_",
                "description": "Generic types for Document:\n*  T - the type of _id\n*  TQueryHelpers - Object with any helpers that should be mixed into the Query type\n*  DocType - the type of the actual Document created"
            },
            "OId": {
                "type": "string",
                "minLength": 24,
                "maxLength": 24,
                "pattern": "[a-f][0-9]"
            },
            "IProfile": {
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/OId"
                    },
                    "customerGroupID": {
                        "$ref": "#/components/schemas/OId"
                    },
                    "type": {
                        "$ref": "#/components/schemas/OId"
                    },
                    "name": {
                        "type": "string"
                    },
                    "officialName": {
                        "type": "string"
                    },
                    "taxNumber": {
                        "type": "number",
                        "format": "double"
                    },
                    "taxOffice": {
                        "type": "string"
                    },
                    "addressCountry": {
                        "type": "string"
                    },
                    "addressCity": {
                        "type": "string"
                    },
                    "addressDistrict": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "contactName": {
                        "type": "string"
                    },
                    "contactPhone": {
                        "type": "string"
                    },
                    "contactEmail": {
                        "type": "string"
                    },
                    "createdAt": {},
                    "updatedAt": {}
                },
                "required": [
                    "_id",
                    "customerGroupID",
                    "type",
                    "name",
                    "officialName",
                    "taxNumber",
                    "taxOffice",
                    "addressCountry",
                    "addressCity",
                    "addressDistrict",
                    "address",
                    "contactName",
                    "contactPhone",
                    "contactEmail",
                    "createdAt",
                    "updatedAt"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "ObjectId": {
                "type": "string"
            },
...
            "User": {
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/OId"
                    },
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "user"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "createdAt": {},
                    "updatedAt": {}
                },
                "required": [
                    "_id",
                    "role",
                    "name",
                    "surname",
                    "email",
                    "password",
                    "createdAt",
                    "updatedAt"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UserSendParams": {
                "properties": {
                    "_id": {
                        "$ref": "#/components/schemas/OId"
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    }
                },
                "required": [
                    "_id",
                    "name",
                    "surname"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UserPostParams": {
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "admin",
                            "user"
                        ]
                    },
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "role",
                    "name",
                    "surname",
                    "email",
                    "password"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "UserUpdateParams": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "surname": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "surname"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "in": "header",
                "name": "x-auth"
            }
        }
    },
    "info": {
        "title": "deneme",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/profile/{id}": {
            "get": {
                "operationId": "GetUserById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "$ref": "#/components/schemas/IProfile"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "properties": {
                                                "_id": {
                                                    "$ref": "#/components/schemas/ObjectId"
                                                }
                                            },
                                            "required": [
                                                "_id"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Profile"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/OId"
                        }
                    }
                ]
            }
        },
...
        "/users/{id}": {
            "get": {
                "operationId": "GetUserById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "properties": {
                                                "_id": {
                                                    "$ref": "#/components/schemas/ObjectId"
                                                }
                                            },
                                            "required": [
                                                "_id"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/OId"
                        }
                    }
                ]
            },
            "patch": {
                "operationId": "UpdateUserById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "properties": {
                                                "_id": {
                                                    "$ref": "#/components/schemas/ObjectId"
                                                }
                                            },
                                            "required": [
                                                "_id"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/OId"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateParams"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "DeleteUserById",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "$ref": "#/components/schemas/Document_any.any.any_"
                                        },
                                        {
                                            "properties": {
                                                "_id": {
                                                    "$ref": "#/components/schemas/ObjectId"
                                                }
                                            },
                                            "required": [
                                                "_id"
                                            ],
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/OId"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateParams"
                            }
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "operationId": "AddUser",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserSendParams"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Users"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPostParams"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}